udpating deploy-vms

+++ openshift-jumpbox/deploy-vm.sh
@@ -10,10 +10,18 @@ PULL_SECRET=$(yq eval '.openshift_pull_secret' "${ANSIBLE_VAULT_FILE}")
 NET_NAME=qubinet
 VM_NAME=openshift-jumpbox-$(echo $RANDOM | md5sum | head -c 5; echo;)
 IMAGE_NAME=Fedora-Cloud-Base-37-1.7.x86_64.qcow2
+DISK_SIZE=50
 sudo rm -rf kcli-profiles.yml
-sudo cp ~/.kcli/profiles.yml kcli-profiles.yml
-sudo python3 profile_generator/profile_generator.py update_yaml ${VM_NAME} openshift-jumpbox/template.yaml --image ${IMAGE_NAME} --user $USER --user-password ${PASSWORD} --net-name ${NET_NAME}  --offline-token ${OFFLINE_TOKEN}  --pull-secret ${PULL_SECRET}
+if [ -f ~/.kcli/profiles.yml ]; then
+  sudo cp  ~/.kcli/profiles.yml kcli-profiles.yml
+else 
+    sudo mkdir -p ~/.kcli
+    sudo mkdir -p /root/.kcli
+fi
+sudo python3 profile_generator/profile_generator.py update_yaml openshift-jumpbox openshift-jumpbox/template.yaml --image ${IMAGE_NAME} --user fedora --user-password ${PASSWORD} --net-name ${NET_NAME}  --offline-token ${OFFLINE_TOKEN}  --pull-secret ${PULL_SECRET} --disk-size ${DISK_SIZE} 
 cat  kcli-profiles.yml
 ansiblesafe -f "${ANSIBLE_VAULT_FILE}" -o 1
-#sudo cp kcli-profiles.yml ~/.kcli/profiles.yml
-#sudo cp kcli-profiles.yml /root/.kcli/profiles.yml
+sudo cp kcli-profiles.yml ~/.kcli/profiles.yml
+sudo cp kcli-profiles.yml /root/.kcli/profiles.yml
+echo "Creating VM ${VM_NAME}"
+sudo kcli create vm -p openshift-jumpbox ${VM_NAME} --wait
